[codegen] split format/generic groups into modules
